Nearly tests compliant MemorySpace

+++ src/m68k/CPU/Instructions/ABCD.cs
@@ -16,9 +16,9 @@ namespace M68k.CPU.Instructions
                 throw new System.ArgumentNullException(nameof(instructionSet));
             }
 
-            uint baseNumber;
+            int baseNumber;
             IInstruction i;
-            for (uint f = 0; f < 2; f++)
+            for (int f = 0; f < 2; f++)
             {
                 if (f == 0)
                 {
@@ -31,9 +31,9 @@ namespace M68k.CPU.Instructions
                     i = new ABCDReg(this);
                 }
 
-                for (uint d = 0; d < 8; d++)
+                for (int d = 0; d < 8; d++)
                 {
-                    for (uint s = 0; s < 8; s++)
+                    for (int s = 0; s < 8; s++)
                     {
                         instructionSet.AddInstruction(baseNumber + (d << 9) + s, i);
                     }
@@ -41,31 +41,31 @@ namespace M68k.CPU.Instructions
             }
         }
 
-        protected uint AbcdAr(uint opcode)
+        protected int AbcdAr(int opcode)
         {
-            uint sreg = (opcode & 0x07);
-            uint dreg = (opcode >> 9) & 0x07;
+            int sreg = (opcode & 0x07);
+            int dreg = (opcode >> 9) & 0x07;
             cpu.DecrementAddrRegister(sreg, 1);
             cpu.DecrementAddrRegister(dreg, 1);
-            uint s = cpu.ReadMemoryByte(cpu.GetAddrRegisterLong(sreg));
-            uint d = cpu.ReadMemoryByte(cpu.GetAddrRegisterLong(dreg));
-            uint result = DoCalc(s, d);
+            int s = cpu.ReadMemoryByte(cpu.GetAddrRegisterLong(sreg));
+            int d = cpu.ReadMemoryByte(cpu.GetAddrRegisterLong(dreg));
+            int result = DoCalc(s, d);
             cpu.WriteMemoryByte(cpu.GetAddrRegisterLong(dreg), result);
             return 18;
         }
 
-        protected uint AbcdDr(uint opcode)
+        protected int AbcdDr(int opcode)
         {
-            uint sreg = (opcode & 0x07);
-            uint dreg = (opcode >> 9) & 0x07;
-            uint s = cpu.GetDataRegisterByte(sreg);
-            uint d = cpu.GetDataRegisterByte(dreg);
-            uint result = DoCalc(s, d);
+            int sreg = (opcode & 0x07);
+            int dreg = (opcode >> 9) & 0x07;
+            int s = cpu.GetDataRegisterByte(sreg);
+            int d = cpu.GetDataRegisterByte(dreg);
+            int result = DoCalc(s, d);
             cpu.SetDataRegisterByte(dreg, result);
             return 6;
         }
 
-        protected DisassembledInstruction DisassembleOp(uint address, uint opcode, bool dataRegMode)
+        protected DisassembledInstruction DisassembleOp(int address, int opcode, bool dataRegMode)
         {
             DisassembledOperand src;
             DisassembledOperand dst;
@@ -83,10 +83,10 @@ namespace M68k.CPU.Instructions
             return new DisassembledInstruction(address, opcode, "abcd", src, dst);
         }
 
-        protected uint DoCalc(uint s, uint d)
+        protected int DoCalc(int s, int d)
         {
             int x = (cpu.IsFlagSet(cpu.XFlag) ? 1 : 0);
-            uint c;
+            int c;
             var lo = (s & 0x0f) + (d & 0x0f) + x;
             if (lo > 9)
             {
@@ -98,7 +98,7 @@ namespace M68k.CPU.Instructions
                 c = 0;
             }
 
-            uint hi = ((s >> 4) & 0x0f) + ((d >> 4) & 0x0f) + c;
+            int hi = ((s >> 4) & 0x0f) + ((d >> 4) & 0x0f) + c;
             if (hi > 9)
             {
                 hi -= 10;
@@ -124,7 +124,7 @@ namespace M68k.CPU.Instructions
                 cpu.ClrFlags(cpu.ZFlag);
             }
 
-            return (uint)result;
+            return result;
         }
 
         private sealed class ABCDData : IInstruction
@@ -136,12 +136,12 @@ namespace M68k.CPU.Instructions
                 this.parent = parent;
             }
 
-            public DisassembledInstruction Disassemble(uint address, uint opcode)
+            public DisassembledInstruction Disassemble(int address, int opcode)
             {
                 return parent.DisassembleOp(address, opcode, true);
             }
 
-            public uint Execute(uint opcode)
+            public int Execute(int opcode)
             {
                 return parent.AbcdDr(opcode);
             }
@@ -156,12 +156,12 @@ namespace M68k.CPU.Instructions
                 this.parent = parent;
             }
 
-            public DisassembledInstruction Disassemble(uint address, uint opcode)
+            public DisassembledInstruction Disassemble(int address, int opcode)
             {
                 return parent.DisassembleOp(address, opcode, false);
             }
 
-            public uint Execute(uint opcode)
+            public int Execute(int opcode)
             {
                 return parent.AbcdAr(opcode);
             }
